requirements.txt
tweepy
requests
python-dotenv
textblob
matplotlib
streamlit
pandas
google-generativeai
